refactor identity resolver to reduce duplication
