+++ src/idex/identity-resolver-nocache.js
@@ -1,101 +1,11 @@
-import { toParams } from '../utils/url'
-import { fromError } from '../utils/emitter'
-import { asParamOrEmpty, asStringParamWhen, asStringParam, mapAsParams } from '../utils/types'
-import { DEFAULT_IDEX_AJAX_TIMEOUT, DEFAULT_IDEX_URL, DEFAULT_REQUESTED_ATTRIBUTES } from '../utils/consts'
-
-function _enrichUnifiedId (response) {
-  if (response && response.nonId && !response.unifiedId) {
-    response.unifiedId = response.nonId
-    return response
-  } else {
-    return response
-  }
-}
-
-function _responseReceived (successCallback) {
-  return response => {
-    let responseObj = {}
-    if (response) {
-      try {
-        responseObj = _enrichUnifiedId(JSON.parse(response))
-      } catch (ex) {
-        fromError('IdentityResolverParser', ex)
-      }
-    }
-    successCallback(responseObj)
-  }
-}
+import { noopCache, makeIdentityResolver } from './common'
 
 /**
  * @param {State} config
  * @param {CallHandler} calls
- * @return {{resolve: function(successCallback: function, errorCallback: function, additionalParams: Object), getUrl: function(additionalParams: Object)}}
+ * @return {IdentityResolver}
  * @constructor
  */
 export function IdentityResolver (config, calls) {
-  try {
-    const nonNullConfig = config || {}
-    const idexConfig = nonNullConfig.identityResolutionConfig || {}
-    const externalIds = nonNullConfig.retrievedIdentifiers || []
-    const source = idexConfig.source || 'unknown'
-    const publisherId = idexConfig.publisherId || 'any'
-    const url = idexConfig.url || DEFAULT_IDEX_URL
-    const timeout = idexConfig.ajaxTimeout || DEFAULT_IDEX_AJAX_TIMEOUT
-    const requestedAttributes = idexConfig.requestedAttributes || DEFAULT_REQUESTED_ATTRIBUTES
-
-    const tuples = []
-    tuples.push(asStringParam('duid', nonNullConfig.peopleVerifiedId))
-    tuples.push(asStringParam('us_privacy', nonNullConfig.usPrivacyString))
-    tuples.push(asParamOrEmpty('gdpr', nonNullConfig.gdprApplies, v => encodeURIComponent(v ? 1 : 0)))
-    tuples.push(asStringParamWhen('n3pc', nonNullConfig.privacyMode ? 1 : 0, v => v === 1))
-    tuples.push(asStringParam('gdpr_consent', nonNullConfig.gdprConsent))
-
-    externalIds.forEach(retrievedIdentifier => {
-      tuples.push(asStringParam(retrievedIdentifier.name, retrievedIdentifier.value))
-    })
-
-    const attributeResolutionAllowed = (attribute) => {
-      if (attribute === 'uid2') {
-        return !config.privacyMode
-      } else {
-        return true
-      }
-    }
-
-    requestedAttributes.filter(attributeResolutionAllowed).forEach(requestedAttribute => {
-      tuples.push(asStringParam('resolve', requestedAttribute))
-    })
-
-    const composeUrl = (additionalParams) => {
-      const originalParams = tuples.slice().concat(mapAsParams(additionalParams))
-      const params = toParams(originalParams)
-      return `${url}/${source}/${publisherId}${params}`
-    }
-    const unsafeResolve = (successCallback, errorCallback, additionalParams) => {
-      calls.ajaxGet(composeUrl(additionalParams), _responseReceived(successCallback), errorCallback, timeout)
-    }
-    return {
-      resolve: (successCallback, errorCallback, additionalParams) => {
-        try {
-          unsafeResolve(successCallback, errorCallback, additionalParams)
-        } catch (e) {
-          errorCallback()
-          fromError('IdentityResolve', e)
-        }
-      },
-      getUrl: (additionalParams) => composeUrl(additionalParams)
-    }
-  } catch (e) {
-    console.error('IdentityResolver', e)
-    fromError('IdentityResolver', e)
-    return {
-      resolve: (successCallback, errorCallback) => {
-        errorCallback()
-        fromError('IdentityResolver.resolve', e)
-      },
-      getUrl: () => {
-        fromError('IdentityResolver.getUrl', e)
-      }
-    }
-  }
+  return makeIdentityResolver(config || {}, calls, noopCache())
 }
